refactor: move types into correct location
